finished manytomany

+++ src/main/java/systema/crm/services/impl/ApplicationRequestServiceImpl.java
@@ -4,11 +4,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import systema.crm.entities.ApplicationRequest;
 import systema.crm.entities.Courses;
+import systema.crm.entities.Operators;
 import systema.crm.repositories.ApplicationRequestRepository;
 import systema.crm.repositories.CoursesRepository;
 import systema.crm.repositories.OperatorsRepository;
 import systema.crm.services.ApplicationRequestService;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -51,6 +53,7 @@ public class ApplicationRequestServiceImpl implements ApplicationRequestService 
 
     @Override
     public void deleteRequest(Long id) {
+        if (isExist(id))
         appRepository.deleteById(id);
     }
 
@@ -88,5 +91,50 @@ public class ApplicationRequestServiceImpl implements ApplicationRequestService 
     @Override
     public Courses getCourse(Long id) {
         return coursesRepository.findById(id).orElse(null);
+    }
+
+    @Override
+    public void assignOperator(Long[] operatorId, Long requestId) {
+        if (isExist(requestId)) {
+            ApplicationRequest request = getRequestById(requestId);
+            List<Operators> operators = request.getOperators();
+            if (operators == null) {
+                operators = new ArrayList<>();
+            }
+
+            for (Long aLong : operatorId) {
+                Operators operator = operatorsRepository.findById(aLong).orElse(null);
+                operators.add(operator);
+            }
+            request.setOperators(operators);
+            request.setHandled(true);
+            saveRequest(request);
+//            request.setOperators(operators);
+//                Operators operator = operatorsRepository.findById(operatorId).orElse(null);
+//                operators.add(operator);
+//                request.setOperators(operators);
+        }
+    }
+
+    @Override
+    public void unAssignOperator(Long operatorId, Long requestId) {
+        if (isExist(requestId)) {
+            ApplicationRequest request = getRequestById(requestId);
+            List<Operators> operators = request.getOperators();
+            if (operators == null) {
+                operators = new ArrayList<>();
+            }
+
+            Operators operator = operatorsRepository.findById(operatorId).orElse(null);
+            operators.remove(operator);
+            request.setOperators(operators);
+            request.setHandled(true);
+            saveRequest(request);
+        }
+    }
+
+    @Override
+    public List<Operators> getOperatorsByRequest(Long requestId) {
+        return getRequestById(requestId).getOperators();
     }
 }
